# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js SugarApp Application API"
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/createdealid": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create Deal ID",
        "description": "",
        "operationId": "createdealid",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dealdataids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deal ID create SuccessFully",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/editterms": {
      "put": {
        "tags": [
          "Merchant"
        ],
        "summary": "This Merchanrt Api is used to edit Deal Term",
        "description": "",
        "operationId": "editterms",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/editdealterm"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/api/updateendtime": {
      "put": {
        "tags": [
          "Merchant"
        ],
        "summary": "This Merchant Api used to Edit DealID end time",
        "description": "",
        "operationId": "updateendtime",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Edittime"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/api/getsubscriptionlist": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "summary": "This is Merchant Api used to get subscription list",
        "description": "",
        "operationId": "getsubscriptionlist",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/scriptionlist"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/api/getredemptionlist": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "summary": "This is Merchant Api used to get Redemption list ",
        "description": "",
        "operationId": "getredemptionlist",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/scriptionlist"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/api/insertimei": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "summary": "This is Merchant Api  used to insert IEMI Number",
        "description": "",
        "operationId": "insertimei",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IMEI"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/api/createjyfdealid": {
      "post": {
        "tags": [
          "Merchant"
        ],
        "summary": "This is Merchant Api  used to create JYF DealID",
        "description": "",
        "operationId": "createjyfdealid",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createjyfdealid"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dealdataids": {
      "required": [
        "dealID"
      ],
      "properties": {
        "dealID": {
          "type": "string",
          "uniqueItems": true
        },
        "dealinfo": {
          "type":"string"
        },
        "dealterms": {
          "type":"string"
        },
        "setdate": {
          "type":"string"
        },
        "settime": {
          "type":"string"
        },
        "endtime": {
          "type":"string"
        }
      }
    },
...
    "editdealterm": {
      "required": [
        "dealID"
      ],
      "properties": {
        "dealID": {
          "type": "string",
          "uniqueItems": true
        },
         "dealterms": {
          "type": "string"
        }
      }
    },
    "Edittime":{
      "required": [
        "dealID"
      ],
      "properties": {
        "dealID": {
          "type": "string",
          "uniqueItems": true
        },
         "endtime": {
          "type": "string"
        }
      }
    },
    "scriptionlist":{
      "required": [
        "dealID"
      ],
      "properties": {
        "dealID": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
    "IMEI":{
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "imei": {
          "type": "array"
        }
      }
    },
    "createjyfdealid":{
       "required": [
        "dealID"
      ],
      "properties": {
        "dealID": {
          "type": "string",
          "uniqueItems": true
        },
        "dealinfo": {
          "type":"string"
        },
        "dealterms": {
          "type":"string"
        },
        "enddate": {
          "type":"string"
        },
        "vendorcustomerlist": {
          "type":"array"
        }
      }
    }
  }
}